Added Focus Punch AI option
This commit adds the option to toggle between the new ai for focus punch
Which checks whether the previous move used by the player was an
attacking move. This was done to make it so you couldn't spam, for
example, Blissey Seismic Toss against opposing FP Swampert. In the
Vanilla AI there is no check for any damaging moves used previously
that could break focus punch, meaning you could just spam an attacking
move to cheese the AI if it saw focus punch did the most damage.

In addition I setup the new get_ai_option battle ai script such that it
should be scalable to any new modifications anyone might want to make
to the ai in the future. Check the new /include/ai_options.h and
/include/constants/ai_options.h for details on how to do this.